Adding Full Backend Changes to BackOffice

+++ EcoBazzar/DataModel/Product.cs
@@ -11,7 +11,7 @@
         public double Rating { get; set; }
         public double Weight { get; set; }
         public int Stock { get; set; }
-        public double? Discount { get; set; }
+        public int? Discount { get; set; }
         public DateTime DateAdded { get; set; }
         public DateTime? DateModified { get; set; }
         public SubCategory SubCategory { get; set; }
